added cpu-stress that tests timer latency
invoke:
	$ ./cpu-stress sec usec
example:

$ cpu-stress 1 10
398781499
1.000010,1.000015,0.000005
Asked delta: 1.000010 seconds
Got delta:   1.000015 seconds
Latency:     0.000005 seconds

First line is the count, a measure of amount of work done by
the cpu-stress program in while(1) count++;

a,b,c represent the asked timer delta, actual timer delta and
the latency respectively.  You ask for a 1s+10us, get an alarm at 1s+15us
and the latency (mostly due to the extra gettimeofday call).

Last 3 lines are output to stderr, so it will not interfere with
logging.